refactor(api): get user id from authorization header

+++ src/api/controllers/assistant_collection_controller.cc
@@ -4,15 +4,17 @@
 
 #include <crow.h>
 #include <nlohmann/json.hpp>
-#include <chrono>
-#include <iomanip>
-#include <sstream>
 #include <string>
 
 using json = nlohmann::json;
 
 namespace controllers {
     crow::response create_assistant_collection(const crow::request &req) {
+        auto auth_header = req.get_header_value("Authorization");
+        if (auth_header.empty()) {
+            return crow::response(401, "No Authorization header provided");
+        }
+
         try {
             json request_body = json::parse(req.body);
 
@@ -20,7 +22,9 @@ namespace controllers {
                 return crow::response(400, "Invalid JSON");
             }
 
-            std::string user_id = request_body.value("user_id", "");
+            auto get_user_id = get_user_id_from_token(auth_header);
+            std::string user_id = *get_user_id;
+
             if (user_id.empty()) {
                 return crow::response(400, "User ID is required");
             }
@@ -49,10 +53,16 @@ namespace controllers {
     }
 
     crow::response get_assistant_collections(const crow::request &req) {
+        auto auth_header = req.get_header_value("Authorization");
+        if (auth_header.empty()) {
+            return crow::response(401, "No Authorization header provided");
+        }
+
         try {
             json request_body = json::parse(req.body);
 
-            std::string user_id = request_body.value("user_id", "");
+            auto get_user_id = get_user_id_from_token(auth_header);
+            std::string user_id = *get_user_id;
 
             if (user_id.empty()) {
                 return crow::response(400, "User ID must be provided");
@@ -78,6 +88,11 @@ namespace controllers {
     }
 
     crow::response delete_assistant_collection(const crow::request &req, const int &assistant_id, const int &collection_id) {
+        auto auth_header = req.get_header_value("Authorization");
+        if (auth_header.empty()) {
+            return crow::response(401, "No Authorization header provided");
+        }
+
         try {
             if (models::AssistantCollection::delete_assistant_collection(assistant_id, collection_id)) {
                 return crow::response(204, "");
@@ -95,6 +110,11 @@ namespace controllers {
     }
 
     crow::response update_assistant_collection(const crow::request &req, const int &assistant_id, const int &collection_id) {
+        auto auth_header = req.get_header_value("Authorization");
+        if (auth_header.empty()) {
+            return crow::response(401, "No Authorization header provided");
+        }
+
         try {
             json request_body = json::parse(req.body);
 
@@ -104,7 +124,8 @@ namespace controllers {
                 return crow::response(404, "Assistant collection not found");
             }
 
-            assistant_collection.user_id = request_body.value("user_id", assistant_collection.user_id);
+            auto get_user_id = get_user_id_from_token(auth_header);
+            assistant_collection.user_id = *get_user_id;
             assistant_collection.updated_at = get_current_time();
 
             if (models::AssistantCollection::update_assistant_collection(assistant_id, collection_id, assistant_collection)) {
@@ -123,6 +144,11 @@ namespace controllers {
     }
 
     crow::response get_assistant_collection_by_id(const crow::request &req, const int &assistant_id, const int &collection_id) {
+        auto auth_header = req.get_header_value("Authorization");
+        if (auth_header.empty()) {
+            return crow::response(401, "No Authorization header provided");
+        }
+
         UserAssistantCollection assistant_collection = models::AssistantCollection::get_assistant_collection_by_id(assistant_id, collection_id);
 
         if (assistant_collection.user_id.empty()) {
