+++ src/api/controllers/collection_controller.cc
@@ -4,23 +4,26 @@
 
 #include <crow.h>
 #include <nlohmann/json.hpp>
-#include <chrono>
-#include <iomanip>
-#include <sstream>
 #include <string>
 
 using json = nlohmann::json;
 
 namespace controllers {
     crow::response create_collection(const crow::request &req) {
+        auto auth_header = req.get_header_value("Authorization");
+        if (auth_header.empty()) {
+            return crow::response(401, "No Authorization header provided");
+        }
+
         try {
             json request_body = json::parse(req.body);
 
             if (!request_body.is_object()) {
                 return crow::response(400, "Invalid JSON");
             }
+            auto get_user_id = get_user_id_from_token(auth_header);
+            std::string user_id = *get_user_id;
 
-            std::string user_id = request_body.value("user_id", "");
             if (user_id.empty()) {
                 return crow::response(400, "User ID is required");
             }
@@ -54,10 +57,16 @@ namespace controllers {
     }
 
     crow::response get_collections(const crow::request &req) {
+        auto auth_header = req.get_header_value("Authorization");
+        if (auth_header.empty()) {
+            return crow::response(401, "No Authorization header provided");
+        }
+
         try {
             json request_body = json::parse(req.body);
 
-            std::string user_id = request_body.value("user_id", "");
+            auto get_user_id = get_user_id_from_token(auth_header);
+            std::string user_id = *get_user_id;
 
             if (user_id.empty()) {
                 return crow::response(400, "User ID must be provided");
@@ -83,6 +92,11 @@ namespace controllers {
     }
 
     crow::response delete_collection(const crow::request &req, const int &collection_id) {
+        auto auth_header = req.get_header_value("Authorization");
+        if (auth_header.empty()) {
+            return crow::response(401, "No Authorization header provided");
+        }
+
         try {
             if (models::Collection::delete_collection(collection_id)) {
                 return crow::response(204, "");
@@ -100,6 +114,11 @@ namespace controllers {
     }
 
     crow::response update_collection(const crow::request &req, const int &collection_id) {
+        auto auth_header = req.get_header_value("Authorization");
+        if (auth_header.empty()) {
+            return crow::response(401, "No Authorization header provided");
+        }
+
         try {
             json request_body = json::parse(req.body);
 
@@ -131,6 +150,11 @@ namespace controllers {
     }
 
     crow::response get_collection_by_id(const crow::request &req, const int &collection_id) {
+        auto auth_header = req.get_header_value("Authorization");
+        if (auth_header.empty()) {
+            return crow::response(401, "No Authorization header provided");
+        }
+
         UserCollection collection = models::Collection::get_collection_by_id(collection_id);
 
         if (collection.user_id.empty()) {
